Fix : Update PassEntity

+++ src/main/java/com/dev/passbatch/repository/pass/PassEntity.java
@@ -17,7 +17,7 @@ public class PassEntity extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long passSeq;
+    private Integer passSeq;
 
     private Integer packageSeq;
